requirements.txt
prompt-toolkit==3.0.28
pyserial==3.5
questionary==1.10.0
wcwidth==0.2.5
